+++ crates/senmei-ml/src/bin/convert.rs
@@ -51,7 +51,7 @@ fn main() -> senmei_ml::Result<()> {
     if input.extension().and_then(|e| e.to_str()) == Some("onnx") {
         senmei_ml::convert_onnx_to_bpk(&args[1], input, out, scale, num_block, shuffle)?;
     } else if input.extension().and_then(|e| e.to_str()) == Some("safetensors") {
-        senmei_ml::convert_safetensors_to_bpk(&args[1], input, out, scale)?;
+        senmei_ml::convert_safetensors_to_bpk(&args[1], input, out, scale, num_block)?;
     } else {
         senmei_ml::convert_pth_to_bpk(&senmei_ml::ConvertOptions {
             arch: args[1].as_str(),